+++ src/components/layout/Footer.jsx
@@ -17,11 +17,11 @@ const Footer = styled.footer`
   margin: 4em ${(props) => props.theme.layout.marginMobile} 2em;
   text-align: center;
 
-  @media screen and ${(props) => props.theme.devices.mobileL} {
+  @media screen and (${(props) => props.theme.devices.mobileL}) {
     margin: 4em ${(props) => props.theme.layout.marginTablet} 2em;
   }
 
-  @media screen and ${(props) => props.theme.devices.tablet} {
+  @media screen and (${(props) => props.theme.devices.tablet}) {
     margin: 4em ${(props) => props.theme.layout.marginDesktop} 2em;
   }
 `;
@@ -33,11 +33,11 @@ const Div = styled.div`
 const SocialMediaTitle = styled.h3`
   margin: 2rem 0 0;
 
-  @media screen and ${(props) => props.theme.devices.mobileL} {
+  @media screen and (${(props) => props.theme.devices.mobileL}) {
     margin: 3rem 0 0;
   }
 
-  @media screen and ${(props) => props.theme.devices.tablet} {
+  @media screen and (${(props) => props.theme.devices.tablet}) {
     margin: 4rem 0 0;
   }
 `;