+++ src/components/layout/Layout.jsx
@@ -19,12 +19,12 @@ export default (props) => {
     font-size: 2rem;
     text-align: center;
 
-    @media screen and ${(props) => props.theme.devices.mobileL} {
+    @media screen and (${(props) => props.theme.devices.mobileL}) {
       margin: 2em ${(props) => props.theme.layout.marginTablet};
       font-size: 3rem;
     }
 
-    @media screen and ${(props) => props.theme.devices.tablet} {
+    @media screen and (${(props) => props.theme.devices.tablet}) {
       margin: 2em ${(props) => props.theme.layout.marginTablet};
       font-size: 3rem;
     }